Working on base class for game pieces and player control.
Added Ball class which has basic routines for accelerating and moving a
ball. Added PlayerInput struct as an abstract representation of what the
player's "controller" is at a given moment in time. Added and event
handling code for setting the player's desired movement.

Currently, running will generate window and run a simple "game" where
the player (a square) can be accelerated with the arrow keys to a max
velocity.